+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 ## Unreleased
 
+- remove HeroUI Pro components (#557)
+- refactor UI components to allow modifications (#579)
+- Add setup method to ChatInterface (#586)
+- Rebuild UI with new dependencies (#589)
+
+## 0.19.1 (2025-05-27)
+
+### Changed
+
+- ragbits-core updated to version v0.19.1
+
+- fix: dont import all persistence strategies in base file (#584)
+
+## 0.19.0 (2025-05-27)
+
+### Changed
+
+- ragbits-core updated to version v0.19.0
+
+- Add persistance component to save chat interactions from ragbits-chat (#556)
+- Add conversation_id parameter to chat interface context (#556)
+- Add uvicorn to dependencies (#578)
+- Remove HeroUI Pro components (#557)
+
 ## 0.18.0 (2025-05-22)
 
 ### Changed